helloworld/pyproject.toml
[project]
name = "helloworld"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "“fistasolutions”", email = "fistasolutions@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.3",
    "python-dotenv>=1.0.1",
]

[project.scripts]
helloworld = "helloworld:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
